Module Project.ExternalStatusCheck

External Status Checks API. https://docs.gitlab.com/ee/api/status_checks.html

val list_for_merge_request : token:Token.t -> project_id:int -> merge_request_iid:string -> unit -> Gitlab_t.status_checks Response.t Monad.t

list_for_merge_request ~project_id ~merge_request_iid For a single merge request, list the external status checks that apply to it and their status.

See List status checks for a merge request.

val set_status : token:Token.t -> project_id:int -> merge_request_iid:string -> sha:string -> external_status_check_id:string -> unit -> Gitlab_t.external_status_check Response.t Monad.t

set_status ~project_id ~merge_request_iid ~sha ~external_status_check_id For a single merge request, use the API to inform GitLab that a merge request has passed a check by an external service.

See Set status of an external status check.

val checks : token:Token.t -> project_id:int -> unit -> Gitlab_t.external_status_checks Response.t Monad.t

checks ~project_id request project's external status checks.

See Get project external status checks.

val create : token:Token.t -> project_id:int -> name:string -> external_url:string -> ?protected_branch_ids:int list -> unit -> Gitlab_t.external_status_check Response.t Monad.t

create create a new external status check for a project.

See Create external status check.

val delete : token:Token.t -> project_id:int -> rule_id:int -> unit -> unit Response.t Monad.t

delete ~project_id ~rule_id an external status check for a project.

See Delete external status check.

val update : token:Token.t -> project_id:int -> rule_id:int -> ?name:string -> ?external_url:string -> ?protected_branch_ids:int list -> unit -> Gitlab_t.external_status_check Response.t Monad.t

update ~project_id ~rule_id an external status check for a project.

See Update external status check.